chore(games): reworks userunique.py to an abstract class with lashprompt.py providing the concrete implementation
