+++ src/interactive/__init__.py
@@ -6,5 +6,5 @@ from interactive.message import MessageInteraction
 from interactive.choice import ChoiceInteraction
 from interactive.messagechoice import MessageChoiceInteraction
 
-from interactive.userunique import UserUniqueView
+from interactive.lashprompt import LashGetPromptView
 from interactive.gatherplayers import GatherPlayersView
